requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
Pillow>=9.0.0
numpy>=1.23.0
opencv-python-headless>=4.7.0
